+++ src/components/__tests__/CheckBox.test.js
@@ -1,4 +1,4 @@
-import CheckBox from "../CheckBox";
+import CheckBox from "../Checkbox";
 
 describe("CheckBox component", () => {
 
@@ -10,7 +10,7 @@ describe("CheckBox component", () => {
 
     it("Returns a checkbox input with a label", () => {
 
-        render(CheckBox({name: "option_1", text: "value_1", value: "value_1"}));
+        render(CheckBox({name: "option_1", text: "value_1", value: "value_1"}).element);
 
         expect(document.querySelector("label")).not.toBeNull();
         expect(document.querySelector("input[type='checkbox']")).not.toBeNull();
@@ -18,42 +18,42 @@ describe("CheckBox component", () => {
 
     it("Label uses the text from the text prop", () => {
 
-        render(CheckBox({name: "option_1", text: "value_1", value: "value_1"}));
+        render(CheckBox({name: "option_1", text: "value_1", value: "value_1"}).element);
 
         expect(document.querySelector("label").textContent).toBe("value_1");
     });
 
     it("checkbox uses the value from the value prop", () => {
 
-        render(CheckBox({name: "option_1", text: "value_1", value: "value_1"}));
+        render(CheckBox({name: "option_1", text: "value_1", value: "value_1"}).element);
 
         expect(document.querySelector("input[type='checkbox']").value).toBe("value_1");
     });
 
     it("checkbox uses the name from the name prop", () => {
 
-        render(CheckBox({name: "option_1", text: "value_1", value: "value_1"}));
+        render(CheckBox({name: "option_1", text: "value_1", value: "value_1"}).element);
 
         expect(document.querySelector("input[type='checkbox']").name).toBe("option_1");
     });
 
     it("Checks checkbox when checked prop is true", () => {
 
-        render(CheckBox({name: "option_1", text: "value_1", value: "value_1", checked: true}));
+        render(CheckBox({name: "option_1", text: "value_1", value: "value_1", checked: true}).element);
 
         expect(document.querySelector("input[type='checkbox']").checked).toBe(true);
     });
 
     it("Unchecks checkbox when checked prop is true", () => {
 
-        render(CheckBox({name: "option_1", text: "value_1", value: "value_1", checked: false}));
+        render(CheckBox({name: "option_1", text: "value_1", value: "value_1", checked: false}).element);
 
         expect(document.querySelector("input[type='checkbox']").checked).toBe(false);
     });
 
     it("Checkbox is checked when label is clicked", () => {
 
-        render(CheckBox({name: "option_1", text: "value_1", value: "value_1"}));
+        render(CheckBox({name: "option_1", text: "value_1", value: "value_1"}).element);
 
         const checkBox = document.querySelector("input[type='checkbox']");
         const label = document.querySelector("label");
@@ -69,7 +69,7 @@ describe("CheckBox component", () => {
 
         const onChange = jest.fn();
 
-        render(CheckBox({name: "option_1", text: "value_1", value: "value_1", onChange: onChange}));
+        render(CheckBox({name: "option_1", text: "value_1", value: "value_1", onChange: onChange}).element);
 
         const checkBox = document.querySelector("input[type='checkbox']");
         checkBox.click();
@@ -81,7 +81,7 @@ describe("CheckBox component", () => {
 
         const onChange = jest.fn();
 
-        render(CheckBox({name: "option_1", text: "value_1", value: "value_1", onChange: onChange}));
+        render(CheckBox({name: "option_1", text: "value_1", value: "value_1", onChange: onChange}).element);
 
         const checkBox = document.querySelector("input[type='checkbox']");
         checkBox.click();